my brain hurts
The motor is now set up to slow down at a linear rate relative to distance from end point, with the slowdown starting at a pre-set point that can be adjusted.

Need to made adjustments so that the constants are accessible from shuffleboard.

A few "project errors" that seem to be non-existint. Worth looking further into.